pyproject.toml
[project]
name = "astrbot-plugin-kahunabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.30.0",
    "cachetools>=6.2.2",
    "colorlog>=6.10.1",
    "httpx>=0.28.1",
    "imgkit>=1.2.3",
    "jinja2>=3.1.6",
    "neo4j>=6.0.3",
    "networkx>=3.5",
    "oauthlib>=3.3.1",
    "peewee>=3.18.3",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "pyppeteer>=2.0.0",
    "quart>=0.20.0",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "sqlalchemy>=2.0.44",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
    "tomli-w>=1.0.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "watchfiles>=1.1.1",
    "werkzeug>=3.1.3",
    "aiocache>=0.12.3",
]

[[tool.uv.index]]
url="http://mirrors.aliyun.com/pypi/simple/"
default=true
